==> src/00_reference_data/03_clustering_ref_data.Rmd <==
---
title: "Clustering reference data"
date: "`r format(Sys.time(), '%d %B, %Y')`"
format:
  html:
    embed-resources: true
    code-fold: show
params:
  fig.path: "./Figures/03/"
editor_options: 
  chunk_output_type: console
---

```{r background-job, eval=FALSE, include=FALSE}
source("../../bin/render_with_jobs.R")

# quarto
# render_html_with_job(out_dir = lab_dir)
# fs::file_move(path = file, new_path = paste0(lab_dir, file))

# currently using quarto for github and kniter for html due to source code option 
render_git_with_job(fig_path = "./Figures/03/")
# Change the figure path from ./Figures/03/ to ../Figures/03/:
system2(command = "sed", stdout = TRUE,
        args = c("-i", "''","-e", 's/src=\\"\\./src=\\"\\.\\./g',
                 paste0("./md_files/", basename("./03_clustering_ref_data.md"))))

# kniter
knit_html_with_job(out_dir = "../../lab_book/reference_data/03_clustering_reference_data", fig_path = "./Figures/03/")
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(
    fig.width   = 6.6929133858,
    fig.path    = params$fig.path,
    dev         = c("png"),
    dpi         = 300,
    fig.align   = "center",
    message     = FALSE,
    warning     = FALSE,
    fig.process = function(filename){
      new_filename <- stringr::str_remove(string = filename, 
                                        pattern = "-1")
      fs::file_move(path = filename, new_path = new_filename)
      ifelse(fs::file_exists(new_filename), new_filename, filename)}) 
# setwd("/Users/vilkal/work/Brolidens_work/Projects/Spatial_DMPA/src/00_reference_data")
```

### Load data and libraries
```{r Load-Library-and-data}
##################
# LOAD LIBRARIES #
##################
library(tidyverse)
library(Seurat)
library(tidyseurat)
library(cowplot)
library(harmony)

source("../../bin/plotting_functions.R")

#########
# PATHS #
#########
input_dir <- "../../results/02_integrate_ref_data/"
result_dir <- "../../results/03_clustering_ref_data/"
if( isFALSE(dir.exists(result_dir)) ) { dir.create(result_dir,recursive = TRUE) }

#############
# LODA DATA #
#############
seuratObj <- readRDS(paste0(input_dir,"seuratObj_harmony.RDS"))

```

### Clustering
```{r Clustering_resolution}
##################################
# EVALUATE CLUSTERING RESOLUTION #
##################################
seuratObj <- FindNeighbors(seuratObj, reduction = "harmony", dims = 1:20, k.param = 20, prune.SNN = 1/15) 

# Clustering with louvain (algorithm 1) or leiden (algorithm 4)
for (res in c(0.1, 0.25, 0.5, 1, 1.5, 2)) {
    seuratObj <- FindClusters(seuratObj, resolution = res, algorithm = 1)
}

# each time you run clustering, the data is stored in meta data columns:
# seurat_clusters - lastest results only CCA_snn_res.XX - for each different
# resolution you test.
```

### Plot clustering resolution
```{r 03a_plot_resolution, fig.width=10, fig.height=5}
# dev.new(width=10, height=5, noRStudioGD = TRUE)
res <- c("RNA_snn_res.1", "RNA_snn_res.1.5")
p <- map(res, ~plot_clusters.fun(seuratObj, cluster=.x, red = "umap_harmony", color = "Brew_all"))
plot_grid(ncol = 2, 
         plotlist = p)


seuratObj <- seuratObj %>%
  rename(Clusters="RNA_snn_res.1.5") %>%
  SetIdent(., value = "Clusters") %>%
  select(-starts_with("RNA_snn_res."), -cell_id) %>%
  select(.cell, sample_name, Clusters, everything())
```

### Save seurat object
```{r save-SeuratObject}
##################################
# SAVE INTERMEDIATE SEURAT OJECT #
##################################
saveRDS(seuratObj, paste0(result_dir,"seuratObj_clustered.RDS"))
# seuratObj<- readRDS(paste0(result_dir,"seuratObj_clustered.RDS"))
```

## Session info
```{r}
sessionInfo()
```

```{r Convert-to-anndata, eval=FALSE, include=FALSE}
library(SeuratDisk)
# seuratObj <- readRDS(paste0(result_dir,"seuratObj_Li_Y_harmony.RDS"))
SaveH5Seurat(seuratObj, filename = paste0(result_dir,"Li_Y_integrated.h5Seurat"))
Convert(paste0(result_dir,"Li_Y_integrated.h5Seurat"), dest = "h5ad")
```

## Paulos Code 
```{r Clustering, eval=FALSE, include=FALSE}
##################################
# EVALUATE CLUSTERING RESOLUTION #
##################################
seuratObj <- FindNeighbors(seuratObj, reduction = "harmony", dims = 1:30, k.param = 20, prune.SNN = 1/15) 

for (res in c(0.1, 0.25, 0.5, 1, 1.5, 2)) {
    seuratObj <- FindClusters(seuratObj, graph.name = "RNA_snn", resolution = res, algorithm = 1)
}

library(niceRplots)
plot_meta(seuratObj, red = "umap_harmony", feat = "RNA_snn_res.1.5", cex = 0.5, label = T)
plot_meta(seuratObj, red = "umap_harmony", feat = "RNA_snn_res.2", cex = 0.5, label = T)

# plot_grid(ncol = 3, DimPlot(seuratObj, reduction = "umap", group.by = "RNA_snn_res.0.5") +
#     ggtitle("leiden_0.5"), DimPlot(seuratObj, reduction = "umap", group.by = "RNA_snn_res.1") +
#     ggtitle("leiden_1"), DimPlot(seuratObj, reduction = "umap", group.by = "RNA_snn_res.2") +
#     ggtitle("leiden_2"))

seuratObj <- SetIdent(seuratObj, value = "RNA_snn_res.1.5")
```
