==> src/00_reference_data/05_subset_celltypes_ref_data.Rmd <==
---
title: "Subsetting cell types"
date: "`r format(Sys.time(), '%d %B, %Y')`"
format:
  html:
    embed-resources: true
    code-fold: show
params:
  fig.path: "./Figures/05/"
  re.run: FALSE
editor_options: 
  chunk_output_type: console
---

```{r background-job, eval=FALSE, include=FALSE}
source("../../bin/render_with_jobs.R")

# quarto
# render_html_with_job(out_dir = lab_dir)
# fs::file_move(path = file, new_path = paste0(lab_dir, file))

# currently using quarto for github and kniter for html due to source code option 
render_git_with_job(fig_path = "./Figures/05/")
# Change the figure path from ./Figures/03/ to ../Figures/03/:
system2(command = "sed", stdout = TRUE,
        args = c("-i", "''","-e", 's/src=\\"\\./src=\\"\\.\\./g',
                 paste0("./md_files/", basename("./04_DGE_ref_data.md"))))

# kniter
knit_html_with_job(out_dir = "../../lab_book/reference_data/05_subset_celltypes_reference_data", fig_path = "./Figures/05/")
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(
    fig.width   = 6.6929133858,
    fig.path    = params$fig.path,
    dev         = c("png"),
    dpi         = 300,
    fig.align   = "center",
    message     = FALSE,
    warning     = FALSE,
    fig.process = function(filename){
      new_filename <- stringr::str_remove(string = filename, 
                                        pattern = "-1")
      fs::file_move(path = filename, new_path = new_filename)
      ifelse(fs::file_exists(new_filename), new_filename, filename)}) 
# setwd("/Users/vilkal/work/Brolidens_work/Projects/Spatial_DMPA/src/00_reference_data")
```

### Load data and libraries
```{r Load-data}
##################
# LOAD LIBRARIES #
##################
library(tidyverse)
library(Seurat)
library(SeuratObject)
library(tidyseurat)
library(harmony)
library(patchwork)
library(openxlsx)

source("../../bin/plotting_functions.R")

#########
# PATHS #
#########
input_dir <- "../../results/03_clustering_ref_data/"
result_dir <- "../../results/05_subset_ref_data/"
marker_dir <- "./marker_genes/"
if( isFALSE(dir.exists(result_dir)) ) { dir.create(result_dir,recursive = TRUE) }
if( isFALSE(dir.exists(marker_dir)) ) { dir.create(marker_dir,recursive = TRUE) }

#############
# LODA DATA #
#############
seuratObj <- readRDS(paste0(input_dir,"seuratObj_clustered.RDS"))
cell_id <- read.xlsx("../../../Spatial_DMPA/data/Reference_data/annotation_single_cell.xlsx", sheet = "edited")

```

### Add cell annotation to seurat object
```{r Add-ST-cell-annotation, eval=params$re.run}
cell_type <- set_names(cell_id$annotation_level1, as.character(cell_id$cluster))
cell_annot_1 <- set_names(cell_id$annotation_level2, as.character(cell_id$cluster))
cell_annot_2 <- set_names(cell_id$annotation_level3, as.character(cell_id$cluster))

# tibble(cell_type)

seuratObj <- seuratObj %>%
  select(-any_of(c("cell_id", "CD8A", "SFRP2", "CD3E", "seurat_clusters"))) %>%
  mutate(., cell_type = cell_type[as.character(.$Clusters)]) %>%
  mutate(., cell_annot_1 = cell_annot_1[as.character(.$Clusters)]) %>%
  mutate(., cell_annot_2 = cell_annot_2[as.character(.$Clusters)]) %>%
  select(.cell, sample_name, Clusters:cell_annot_2, everything())

```

### Subset SeuratObj
```{r subset_cells, eval=params$re.run}
seuratObj_nested <- seuratObj %>%
  filter(., grepl("fibroblasts|endothelial|epithelial|immune", .$cell_type)) %>%
  filter(., ifelse(!(cell_type=="immune" & cell_annot_1 != "T cell"), TRUE, FALSE)) %>%
  nest(data = -cell_type)
```

```{r Find_HVG, eval=FALSE, include=FALSE}
################################
# SPLIT INTO SEPERATE DATASETS #
################################
seuratObj_nested <- seuratObj %>%
  mutate(subset = cell_type) %>% 
  nest(data = -cell_type)%>%
  mutate(data = imap(
    data, ~ .x %>%
      #NormalizeData(., normalization.method = "LogNormalize",verbose = FALSE) %>%
      FindVariableFeatures(selection.method = "vst", 
                           nfeatures = 2000, 
                           verbose = FALSE) )) %>%
  mutate(data = setNames(.[["data"]], .$cell_type))

#########################################
# FIND HIGLY VARIABLE GENES PER SUBSET #
#########################################
hig_var <- seuratObj_nested %>%
  .$data %>%
  map(., ~ unique(unlist(.x@assays$RNA@var.features))) %>%
  # remove all VDJ-genes from list of HVG:
  map(., ~ setdiff(.x, str_subset(.x, "^IGH|^IGK|^IGL|^TRA|^TRB|^TRD|^TRG")))


hig_var.fun <- function(Obj){
  hig_var <- Obj@assays$RNA@var.features %>%
    # remove all VDJ-genes from list of HVG:
    setdiff(., str_subset(., "^IGH|^IGK|^IGL|^TRA|^TRB|^TRD|^TRG"))
  Obj@assays$RNA@var.features <- hig_var
  return(Obj)
}


###########################
# DIMENTIONAITY REDUCTION #
###########################
seuratObj_nested <- seuratObj_nested %>%
  .$data %>%
  map2(., hig_var, ~ ScaleData(.x, verbose = FALSE, features = .y )) %>%
  map2(.,  ~ RunPCA(verbose = FALSE, npcs = 50)) %>%
  map2(.,  ~ RunUMAP(dims = 1:50, n.components = 2L) )

# seuratObj <- seuratObj_nested$data$fibroblasts %>%
#   ScaleData(verbose = FALSE, features = hig_var$fibroblasts ) %>%
#   RunPCA(verbose = FALSE, npcs = 50) %>%
#   RunUMAP(dims = 1:50, n.components = 2L) 
```

## Identify Highly Variable Genes (HVG) in each subset
```{r HVG, eval=params$re.run}
#########################################
# FIND HIGLY VARIABLE GENES PER SUBSET #
#########################################
hig_var.fun <- function(Obj){
  hig_var <- Obj@assays$RNA@var.features %>%
    # remove all VDJ-genes from list of HVG:
    setdiff(., str_subset(., "^IGH|^IGK|^IGL|^TRA|^TRB|^TRD|^TRG"))
  Obj@assays$RNA@var.features <- hig_var
  return(Obj)
}

###########################
# DIMENTIONAITY REDUCTION #
###########################
seuratObj_nested <- seuratObj_nested %>%
  mutate(data = map(
    data, ~ .x %>%
      FindVariableFeatures(verbose = FALSE) %>%
      hig_var.fun(.) %>%
      RunPCA(npcs = 20, verbose = FALSE) %>%
      RunHarmony(group.by.vars = "sample_name", 
             reduction = "pca", 
             dims.use = 1:20, 
             assay.use = "RNA") %>%
      RunUMAP(reduction = "harmony", dims = 1:20, 
              n.components = 2L, verbose = FALSE) %>%
      FindNeighbors(verbose = FALSE, reduction = "harmony", dims = 1:20) )) #%>%
      #FindClusters( verbose = FALSE, resolution = 0.8)

# seuratObj_fibro <- seuratObj_nested$data$fibroblasts %>%
#   FindVariableFeatures(verbose = FALSE) %>%
#   hig_var.fun(.) %>%
#   RunPCA(npcs = 10, verbose = FALSE) %>%
#   FindNeighbors(verbose = FALSE) %>%
#   FindClusters(method = "igraph", verbose = FALSE) %>%
#   RunUMAP(reduction = "pca", dims = 1:10, n.components = 3L, verbose = FALSE)

```

### Load allready saved data
```{r load-nested-seuratObject, include=FALSE, eval=isFALSE(params$re.run)}
# Load data
seuratObj_nested <- readRDS(paste0(result_dir,"subset_celltypes.RDS"))
```

## Clustering
```{r Clustering_resolution}
##################################
# EVALUATE CLUSTERING RESOLUTION #
##################################
# Clustering with louvain (algorithm 1) or leiden (algorithm 4)
clus_res.fun <- function(obj){
  for (res in c(0.1, 0.25, 0.5, 1, 1.5, 2)) {
    obj <- FindClusters(obj, resolution = res, algorithm = 1)
  }
  return(obj)
}

seuratObj_nested <- seuratObj_nested %>%
  mutate(data = map(
    data, ~ .x %>%
      clus_res.fun(.) ))

# each time you run clustering, the data is stored in meta data columns:
# seurat_clusters - lastest results only CCA_snn_res.XX - for each different
# resolution you test.

```

```{r 05a_UMAP_resolution, fig.width=3*4, fig.height=3*5}
# dev.new(width=3*4, height=3*5, noRStudioGD = TRUE) 
res <- c("RNA_snn_res.0.1", "RNA_snn_res.0.25","RNA_snn_res.0.5","RNA_snn_res.1") # , "RNA_snn_res.1.5"
clus_plots <- seuratObj_nested %>%
  add_column(., Res = list(res)) %>%
  rowwise() %>%
  mutate(Plots = list(map2(Res,cell_type, ~plot_clusters.fun(data, cluster=.x, red = "UMAP", title = paste0(.y, "\n", .x)))) )

clus_plots <- clus_plots %>%
  unnest(c("Plots"))

wrap_plots(clus_plots$Plots, ncol=4)
```

```{r choose-cluster-resolution}
res <- c(fibroblasts="RNA_snn_res.0.1", 
         immune="RNA_snn_res.0.25",
         endothelial="RNA_snn_res.0.25", 
         epithelial="RNA_snn_res.0.5",
         granulocytes="RNA_snn_res.0.5"
         )

seuratObj_nested <- seuratObj_nested %>%
  mutate(name = res) %>%
  mutate(data = imap(
    data, ~ .x %>%
      mutate("Clus_{.y}" := factor(unlist(.[[res[.y]]])), .after = "Clusters") %>% 
      #select(.cell, sample_name, Clusters_old="Clusters", Clusters=name[[.y]], everything()) %>%
      #SetIdent(., value = "Clusters") %>%
      select(-starts_with("RNA_snn_res."))))

```

### Save seurat object
```{r save_SeuratObject, eval=params$re.run}
###########################
# SAVE INTERMEDIATE OJECT #
###########################
saveRDS(seuratObj_nested, paste0(result_dir,"subset_celltypes.RDS"))
# seuratObj_nested <- readRDS(paste0(result_dir,"subset_celltypes.RDS"))
```

```{r 05_tidySeurat_plot, eval=FALSE}
seuratObj_nested %>%

  # Convert to tibble otherwise Seurat drops reduced dimensions when unifying data sets.
  mutate(data = map(data, ~ .x %>% as_tibble())) %>%
  unnest(data) %>%

  # Define unique clusters
  unite("cluster", c(cell_type, Clusters), remove = FALSE) %>%

  # Plotting
  ggplot(aes(UMAP_1, UMAP_2, color = cluster)) +
  geom_point() +
  facet_wrap(~cell_type) +
  my_theme
```

```{r 05b_new_clustering, fig.height=12}
# dev.new(width=6.67, height=12, noRStudioGD = TRUE) 
seuratObj_nested <- seuratObj_nested %>%
  mutate(subset = paste0("Clus_", .$cell_type)) %>%
  mutate(Clus_plots = map(data, ~plot_clusters.fun(.x, cluster="Clusters", red = "umapharmony"))) %>%
  mutate(Subset_plots = pmap(., ~plot_clusters.fun(..2, cluster=..4, red = "umapharmony"))) # "umap_harmony"

p <- c(seuratObj_nested$Clus_plots, seuratObj_nested$Subset_plots)

wrap_plots(p, ncol = 2, byrow = FALSE)
```

## Identify marker genes to seperate clusters
```{r Find_marker_genes, eval=params$re.run}
# Compute differential expression one clusters against rest 
seuratObj_nested <- seuratObj_nested[1:4,] %>%
  mutate(marker_genes = map(
    data, ~ .x %>%
    FindAllMarkers(.,
      log2FC.threshold = 0.1, 
      test.use = "wilcox",
      min.pct = 0.1, 
      min.diff.pct = 0.1, 
      only.pos = TRUE, 
      max.cells.per.ident = 50,
      assay = "RNA", slot = "data") ) %>%
      set_names(., seuratObj_nested[1:4,]$cell_type))

# Get difference in expression between one cluster and the rest
seuratObj_nested <- seuratObj_nested %>%
  mutate(marker_genes = map(
    marker_genes, ~ .x %>%
      mutate(pct.diff = -.$pct.2-.$pct.1) %>%
      mutate(log.pct.diff = -log2(.$pct.2/.$pct.1)) ))

marker_genes <- seuratObj_nested %>% select(cell_type, marker_genes)
```

## Save marker genes object
```{r save_marker_genes, eval=params$re.run}
###########################
# SAVE INTERMEDIATE OJECT #
###########################
saveRDS(marker_genes, paste0(result_dir,"marker_genes.RDS"))
# marker_genes <- readRDS(paste0(result_dir,"marker_genes.RDS"))
```

### Load allready saved data
```{r load-marker-genes, include=FALSE, eval=isFALSE(params$re.run)}
# Load data
marker_genes <- readRDS(paste0(result_dir,"marker_genes.RDS"))
```

```{r Get_best_markers}
# Identify the top genes that have a high difference in expression
remove <- "^IGH|^IGK|^IGL|^TRA|^TRB|^TRD|^TRG"
marker_genes <- marker_genes %>%
  mutate(top25 = map(
    marker_genes, ~ .x %>%
      group_by(cluster) %>%
      top_n(-40, p_val_adj) %>%
      top_n(20, pct.diff) %>%
      #top_n(10, log.pct.diff)
      
      ungroup() %>%
      # remove all VDJ-genes from list of HVG:
      filter(., !(.$gene %in% str_subset(.$gene, remove)))
    ))
    
# top25 <- marker_genes$data$fibroblasts %>%
#   group_by(cluster) %>%
#   top_n(-40, p_val_adj) %>%
#   top_n(30, pct.diff) %>%
#   top_n(20, log.pct.diff) 

marker_genes <- marker_genes %>%
  mutate(top25_gr = map(
    top25, ~ .x %>%
    group_by(cluster) %>%
    group_split() %>%
    set_names(., seq(0, length(.)-1) ) ))
  
# top25[grep("NCR1",top25$gene),]
marker_genes
```

```{r 05c_doptplot_marker_genes, fig.width=10, fig.height=length(unique(marker_genes$top25[[1]]$gene))/6+2, eval=TRUE}
# dev.new(width=6, height=length(unique(marker_genes$top25[[1]]$gene))/6+2, noRStudioGD = TRUE) 
##########################
# SELECT AND ORDER GENES #
##########################
# select spesific genes that can be plottet seperatly:
genes_immune <- rownames(seuratObj_nested$data$immune)[grepl( "CD4|CD3|TR[ABGD]C|IL[0-9]|CC[LR][0-9]",
                                                              rownames(seuratObj_nested$data$immune))]

marker_genes <- marker_genes %>%
  mutate(genes = pmap(., ~rev(as.character(unique(..3$gene)))) ) %>%
  mutate(genes_ord = pmap(bind_cols(seuratObj_nested[1:4,c("data")], .), 
                          ~niceRplots::getcluster(..1, ..6, paste0("Clus_",..2)))) %>%
  mutate(genes_ord = pmap(., ~names(sort(..6)) ))

ord <- niceRplots::getcluster(seuratObj_nested$data[[1]], marker_genes$genes[[1]], "Clus_fibroblasts")
#######################################
# DOTPLOT 20 MARKER GENES PER CLUSTER #
#######################################
marker_genes <- marker_genes %>%
  mutate(height = map_dbl(genes, ~length(.x)/6+2) ) %>%
  mutate(dotplots = pmap(bind_cols(seuratObj_nested[1:4,c("data")], .), 
                        ~DotPlot(.x, features = ..7, 
                          group.by = "Clusters", assay = "RNA") + coord_flip()
  )) %>%
  mutate(dotplots_ = pmap(bind_cols(seuratObj_nested[1:4,c("data")], .), 
                        ~DotPlot(..1, features = ..7, 
                          group.by = paste0("Clus_",..2), assay = "RNA") + coord_flip()
  ))

pmap(marker_genes, ~ggsave(paste0("./Figures/05/", "dotplot_Clusters_",..1,".pdf"), 
                           ..8, height = ..7, width = 10, limitsize = FALSE))
pmap(marker_genes, ~ggsave(paste0("./Figures/05/", "dotplot_Clus_",..1,".pdf"), 
                           ..9, height = ..7, width = 10, limitsize = FALSE))

marker_genes$dotplots_$fibroblasts
```

```{r 05d_barplot_top20, fig.width=5*3, fig.height=ceiling(21/5)*3}
# dev.new(width=3, height=4, noRStudioGD = TRUE) 
barplot.fun <- function(top25){
# pdf(paste0("./Figures/05/","top_DEG.pdf") )#, width = 4*10*300, height = 30*4*300#, res = 300
  
  #top25 <- top25 %>% group_by(cluster) %>% slice_head(., n = 10)
  
  clus <- unique(top25[["cluster"]]) %>% set_names(., paste0("Clus ",.))
  #par(mar = c(4, 5, 3, 4)) #b,l,t,r
  p <- imap(clus, ~wrap_elements(panel = 
                          ~barplot(sort(setNames(top25[["avg_log2FC"]], 
                                                 top25[["gene"]])[top25[["cluster"]] == .x], F),
                                   horiz = T, cex.names = 1, las = 1, border = "white", yaxs = "i") +
                            title(paste0(.y, " vs. rest"), adj = 0.2, line = 1) +
                            abline(v = c(0, 0.25), lty = c(1, 2)),
                          clip = TRUE
                ))
  # op <- par(mar = c(2, 4, 2, 4) + 0,
  #           oma = c(0,0,0,0) + 0.1,
  #           bg = "transparent", xpd = NA)
  p <- wrap_plots(p)
  par(op) # seems to not be doing anything when you save by ggsave
  return(p)
# dev.off()
}


marker_genes <- marker_genes %>%
  mutate(barplot = pmap(., ~barplot.fun(..3))) %>%
  mutate(Clus_n = map_dbl(top25, ~length(unique(.x$cluster)) )) %>%
  mutate(barplot = set_names(.$barplot, .$cell_type))

# dev.new(width=ncol*3, height=21/ncol*4, noRStudioGD = TRUE) 
par(mar = c(2, 4, 2, 4) + 0, 
    oma = c(0,3,0,0) + 0.1,
    bg = "transparent", xpd = NA)  #bottom, left, top, right
marker_genes$barplot$fibroblasts + plot_layout(ncol = 5)


n_col <- 5
pmap(marker_genes, ~ggsave(paste0("./Figures/05/",..1,"_top_DEG.pdf"), 
                           ..10 + plot_layout(ncol = n_col),
                           bg = "transparent",
                           height = ceiling(..11/n_col)*6, 
                           width = n_col*4.5, limitsize = FALSE))
```

## Session info
```{r}
sessionInfo()
```

## Paulos Code
```{r eval=FALSE, include=FALSE}
library(niceRplots)
#pdf("out.pdf", width = 4*10, height = 27*4, useDingbats = F)

for(df in top25_gr){
  for(gene in df){
    png(paste0(name,"out_gr.png"), width = 4*10*300, height = 30*4*300, res = 300)
  #par(mfrow=c(30,10), mar=c(1,1,1,1))
    par(mfrow=c(6,4), mar=c(1,1,1,1))
    plot_feat(seuratObj, 
              red = "umap_harmony", 
              feat = gene, cex = 0.5, 
              label = "RNA_snn_res.1.5"
              #, add_legend=T
            )
  }
}
dev.off()

plot_feat(seuratObj, 
              red = "umap_harmony", 
              feat = "CTSK", cex = 0.5, 
              label = "RNA_snn_res.1.5"
              #, add_legend=T
            )
```

```{r eval=FALSE, include=FALSE}
g <- c("CCL2", "COL1A1", "SYTL3", "IGFBP4", "PTGDS", "LTB")
genes <- c("CD8A", "MYOZ2", "CD3E", "EPCAM", "COL6A1", "CD4", "JCHAIN", "KRT13")
Fibroblast <- c("APOD", "DCN")
NK <- c()

library(niceRplots)
#pdf("out.pdf", width = 4*10, height = 27*4, useDingbats = F)
png("out_.png", width = 4*10*300, height = 30*4*300, res = 300)
par(mfrow=c(30,10), mar=c(1,1,1,1))
#par(mfrow=c(2,4), mar=c(1,1,1,1))
for(i in top25$gene[1:200]){
  plot_feat(seuratObj, red = "umap_harmony", feat = i, cex = 0.5, label = "RNA_snn_res.1.5"#, add_legend=T
            )
}
dev.off()

annot <- c(FIB_DCN = "DCN",
           NK_cell = "NCR1",
           )

marker_genes[grep("NCR1",rownames(marker_genes)),]
```
